refactor: move validation logic from resource to individual fields
- Introduced validation() and getValidationRules() in Field via HasValidation trait
- Removed dependency on validationRules() method from BuildoraResource
- Updated store() and update() methods in BuildoraController to dynamically collect rules from fields
