+++ src/Fields/Field.php
@@ -2,39 +2,32 @@
 
 namespace Ginkelsoft\Buildora\Fields;
 
+use Ginkelsoft\Buildora\Fields\Traits\HasSearch;
+use Ginkelsoft\Buildora\Fields\Traits\HasValidation;
+use Ginkelsoft\Buildora\Fields\Traits\HasVisibility;
+use Ginkelsoft\Buildora\Fields\Traits\HasLayout;
+
 /**
  * Class Field
  *
  * Represents a generic form or table field within the Buildora system.
- * Supports visibility toggles, searchability, read-only state, labels, help text, and export behavior.
  */
 class Field
 {
-    protected bool $startNewRow = false;
-    protected array $columnSpan = ['default' => 12];
-    protected bool $isSearchable = false;
-    protected string $searchable;
+    use HasSearch;
+    use HasVisibility;
+    use HasLayout;
+    use HasValidation;
+
     public string $name;
     public string $label;
     public string $type;
-    protected ?string $helpText = null;
     public bool $sortable = false;
     public bool $readonly = false;
     public mixed $displayValue = null;
-
-    /**
-     * Visibility context array:
-     * - table, create, edit, export, detail
-     */
-    public array $visibility = [
-        'table' => true,
-        'create' => true,
-        'edit' => true,
-        'export' => true,
-        'detail' => true,
-    ];
-
     public mixed $value = null;
+
+    protected ?string $helpText = null;
 
     /**
      * Field constructor.
@@ -51,7 +44,7 @@ class Field
     }
 
     /**
-     * Static factory method to create a new field.
+     * Static factory method to create a new field instance.
      *
      * @param string $name
      * @param string|null $label
@@ -64,7 +57,7 @@ class Field
     }
 
     /**
-     * Set a custom label.
+     * Set the display label for the field.
      *
      * @param string $label
      * @return self
@@ -76,7 +69,7 @@ class Field
     }
 
     /**
-     * Enable or disable sorting for this field.
+     * Mark the field as sortable.
      *
      * @param bool $condition
      * @return self
@@ -88,7 +81,7 @@ class Field
     }
 
     /**
-     * Set the field as readonly.
+     * Mark the field as read-only.
      *
      * @param bool $condition
      * @return self
@@ -100,92 +93,7 @@ class Field
     }
 
     /**
-     * Set visibility for a specific context (table, create, edit, etc).
-     *
-     * @param string $context
-     * @param bool $condition
-     * @return self
-     */
-    public function show(string $context, bool $condition = true): self
-    {
-        if (isset($this->visibility[$context])) {
-            $this->visibility[$context] = $condition;
-        }
-        return $this;
-    }
-
-    /**
-     * Hide the field in a given context.
-     *
-     * @param string $context
-     * @return self
-     */
-    public function hide(string $context): self
-    {
-        return $this->show($context, false);
-    }
-
-    /**
-     * Check if the field is visible in a given context.
-     *
-     * @param string $context
-     * @return bool
-     */
-    public function isVisible(string $context): bool
-    {
-        return $this->visibility[$context] ?? false;
-    }
-
-    /**
-     * Enable searching on this field.
-     *
-     * @param string $term
-     * @return static
-     */
-    public function searchable(string $term = ''): static
-    {
-        if (! $this->supportsSearch()) {
-            throw new \LogicException(static::class . " does not support searchable().");
-        }
-
-        $this->isSearchable = true;
-        $this->searchable = $term ?: $this->name;
-
-        return $this;
-    }
-
-    /**
-     * Get the database column used for searching.
-     *
-     * @return string
-     */
-    public function getSearchColumn(): string
-    {
-        return $this->searchable ?? $this->name;
-    }
-
-    /**
-     * Determine whether the field is marked as searchable.
-     *
-     * @return bool
-     */
-    public function isSearchable(): bool
-    {
-        return $this->supportsSearch() && $this->isSearchable;
-    }
-
-    /**
-     * Determine if this field type supports searching.
-     *
-     * @return bool
-     */
-    public function supportsSearch(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Set help text for the field.
+     * Set the help text for the field.
      *
      * @param string $text
      * @return self
@@ -197,7 +105,7 @@ class Field
     }
 
     /**
-     * Get the help text associated with this field.
+     * Get the help text.
      *
      * @return string|null
      */
@@ -206,52 +114,51 @@ class Field
         return $this->helpText;
     }
 
-    public function columnSpan(int|array $value): static
-    {
-        if (is_array($value)) {
-            $this->columnSpan = $value;
-        } else {
-            $this->columnSpan = ['default' => $value];
-        }
-
-        return $this;
-    }
-
-    public function getColumnSpan(): array
-    {
-        return $this->columnSpan;
-    }
-
-
-    public function startNewRow(bool $value = true): static
-    {
-        $this->startNewRow = $value;
-        return $this;
-    }
-
-    public function shouldStartNewRow(): bool
-    {
-        return $this->startNewRow;
-    }
-
+    /**
+     * Get the value to display in a table or detail view.
+     *
+     * @param mixed $model
+     * @return string
+     */
     public function getDisplayValue(mixed $model): string
     {
         $value = $model->{$this->name} ?? null;
-
-        // Leeg? Streepje
-        if (is_null($value)) {
-            return '-';
-        }
-
-        // HTML-escaped als standaard
-        return e($value);
+        return is_null($value) ? '-' : e($value);
     }
 
-
+    /**
+     * Hide the field from the table view.
+     *
+     * @return self
+     */
     public function hideFromTable(): self { return $this->hide('table'); }
+
+    /**
+     * Hide the field from the create form.
+     *
+     * @return self
+     */
     public function hideFromCreate(): self { return $this->hide('create'); }
+
+    /**
+     * Hide the field from the edit form.
+     *
+     * @return self
+     */
     public function hideFromEdit(): self { return $this->hide('edit'); }
+
+    /**
+     * Hide the field from export output.
+     *
+     * @return self
+     */
     public function hideFromExport(): self { return $this->hide('export'); }
+
+    /**
+     * Hide the field from the detail view.
+     *
+     * @return self
+     */
     public function hideFromDetail(): self { return $this->hide('detail'); }
 
     /**
@@ -269,6 +176,10 @@ class Field
             'readonly' => $this->readonly,
             'visibility' => $this->visibility,
             'value' => $this->value,
+            'helpText' => $this->helpText,
+            'columnSpan' => $this->columnSpan,
+            'startNewRow' => $this->startNewRow,
         ];
     }
 }
+
